+++ Assets/UniVRM-1.0/UnityBuilder/VrmLoader.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ef069ccddc286cc4e930ba33447b0094
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
